Modify the software name in the copyright and add UT use cases

+++ include/pmu.h
@@ -1,6 +1,6 @@
 /******************************************************************************
  * Copyright (c) Huawei Technologies Co., Ltd. 2024. All rights reserved.
- * gala-gopher licensed under the Mulan PSL v2.
+ * libkperf licensed under the Mulan PSL v2.
  * You can use this software according to the terms and conditions of the Mulan PSL v2.
  * You may obtain a copy of Mulan PSL v2 at:
  *     http://license.coscl.org.cn/MulanPSL2
@@ -125,7 +125,7 @@ struct PmuData {
     unsigned cpu;                   // cpu id
     struct CpuTopology *cpuTopo;    // cpu topology
     const char *comm;               // process command
-    int period;                     // number of Samples
+    uint64_t period;                     // number of Samples
     uint64_t count;                 // event count. Only available for Counting.
     struct PmuDataExt *ext;         // extension. Only available for Spe.
